feat: Split PWAManager using Main Controller Pattern
- Extract PWAServiceWorkerManager (SW registration, updates, cache)
- Extract PWAInstallationManager (install prompts, UI handling, analytics)
- Refactor main PWAManager as lightweight orchestrator
- Reduce from 2,968 words to 1,022 words (66% reduction)
- Maintain backward compatibility for all public APIs
- Apply error handling and fallback mechanisms
