+++ .kiro/specs/core-file-splitting-phase-e1/tasks.md
@@ -39,27 +39,27 @@
     - Test singleton pattern and configuration system integration
     - _Requirements: 1.1, 2.4, 10.2_
 
-- [ ] 3. PWAManager.js splitting (2,968 words → 3 components)
-  - [ ] 3.1 Create PWAServiceWorkerManager component
+- [x] 3. PWAManager.js splitting (2,968 words → 3 components)
+  - [x] 3.1 Create PWAServiceWorkerManager component
     - Extract Service Worker registration and management methods
     - Implement cache strategy and SW lifecycle management
     - Create unit tests for Service Worker functionality
     - _Requirements: 3.1, 3.2, 10.1_
 
-  - [ ] 3.2 Create PWAInstallationManager component
+  - [x] 3.2 Create PWAInstallationManager component
     - Extract installation prompt and app installation methods
     - Implement installation UI handling and analytics
     - Create unit tests for installation functionality
     - _Requirements: 3.1, 3.2, 10.1_
 
-  - [ ] 3.3 Refactor main PWAManager to use sub-components
+  - [x] 3.3 Refactor main PWAManager to use sub-components
     - Update constructor to initialize sub-components
     - Modify public methods to delegate to appropriate sub-components
     - Maintain existing public API for backward compatibility
     - Implement error handling and fallback mechanisms
     - _Requirements: 3.3, 3.4, 11.1, 11.2_
 
-  - [ ] 3.4 Verify PWAManager split completion
+  - [x] 3.4 Verify PWAManager split completion
     - Check all files are under 2,500 words
     - Run existing tests to ensure PWA functionality works
     - Test all public API methods and integration points
